# core-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для анализа активности студентов.",
        "title": "Student Analytics API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/analytics/{student_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Получить аналитику студента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Студента",
                        "name": "student_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.StudentAnalytics"
                        }
                    }
                }
            }
        },
        "/log": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Отправить лог активности",
                "parameters": [
                    {
                        "description": "Данные лога",
                        "name": "log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.StudentLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.StudentAnalytics": {
            "type": "object",
            "properties": {
                "analyzed_at": {
                    "type": "string"
                },
                "avg_time_per_task": {
                    "type": "number"
                },
                "cluster_group": {
                    "type": "string"
                },
                "engagement_score": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "recommendations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "student_id": {
                    "type": "integer"
                },
                "success_rate": {
                    "type": "number"
                },
                "topic_efficiency": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                }
            }
        },
        "domain.StudentLog": {
            "type": "object",
            "properties": {
                "action_type": {
                    "type": "string"
                },
                "attempts": {
                    "type": "integer"
                },
                "correct": {
                    "type": "boolean"
                },
                "difficulty": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "material_id": {
                    "type": "string"
                },
                "selected_distractor": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                },
                "time_spent_on_mat": {
                    "type": "integer"
                },
                "time_spent_on_question": {
                    "type": "integer"
                },
                "time_spent_sec": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}